Split response and write mappers
